Custom reqhandlers

+++ tests/test_streaming.py
@@ -36,32 +36,36 @@ class MockProxyResource(DAVNonCollection):
     def __init__(self, path, environ, target_path):
         super(MockProxyResource, self).__init__(path, environ)
         self.target_path = target_path
+        self.worker = None
 
     def beginWrite(self, contentType=None):
-        print("beginWrite: {}".format(self.target_path))
+        # print("beginWrite: {}".format(self.target_path))
         queue = FileLikeQueue(maxsize=1)
 
         # Simulate an asynchrounous consumer. We use a file, so we can check 
         # the result from the parent unittest process. In real live this could be
         # requests.post(..., data=queue), ...
         def _consumer():
-            print("_consumer: {}".format(self.target_path))
+            # print("_consumer: {}".format(self.target_path))
             with open(self.target_path, "wb") as f:
-                print("_consumer: read()...")
+                s = 0
+                # print("_consumer: read()...")
                 data = queue.read()
                 while data:
-                    print("_consumer: read(): write")
-                    f.write(data)
+                    s += len(data)
+                    # print("_consumer: read(): write")
+                    f.write(compat.to_bytes(data))
                     data = queue.read()
-            print("_consumer(): done")
+            # print("_consumer(): done", s)
 
-        worker = threading.Thread(target=_consumer)
-        worker.setDaemon(True)
-        worker.start()        
+        self.worker = threading.Thread(target=_consumer)
+        self.worker.setDaemon(True)
+        self.worker.start()        
         return queue
 
     def endWrite(self, withErrors):
         print("endWrite: {}".format(self.target_path))
+        self.worker.join()
 
 
 class MockProxyProvider(DAVProvider):
@@ -93,16 +97,62 @@ class BasicTest(unittest.TestCase):
 
     def tearDown(self):
         os.remove(self.test_file)
-        os.remove(self.target_path)
+        if os.path.isfile(self.target_path):
+            os.remove(self.target_path)
         self.provider = None
+
+    def testFileLikeQueueUnsized(self):
+        # queue of unlimited size
+        q = FileLikeQueue()
+        q.write("*" * 42)
+        q.write("*" * 3)
+        # unsized reads will return chunks as queued
+        res = q.read()
+        self.assertEqual(len(res), 42)
+        res = q.read()
+        self.assertEqual(len(res), 3)
+        q.close()  # subsequent reads will return "" instead of blocking
+        res = q.read()
+        self.assertEqual(res, "", "Read after close() returns ''")
+        # subsequent write will raise
+        self.assertRaises(ValueError, q.write, "***")
+
+    def testFileLikeQueue(self):
+        # queue of unlimited size
+        q = FileLikeQueue()
+        # queue 32 bytes
+        q.write("*" * 7)
+        q.write("*" * 11)
+        q.write("*" * 5)
+        q.write("*" * 9)
+        q.close()
+        # sized reads will return chunks as demanded
+        for i in range(6):
+            self.assertEqual(len(q.read(5)), 5)
+        self.assertEqual(len(q.read(5)), 2, "last chunk delivers the reminder")
+        self.assertEqual(len(q.read(5)), 0, "furtehr read() returns ''")
+        # self.assertEqual(q.size, 0)
+
+    def testFileLikeQueueAll(self):
+        # queue of unlimited size
+        q = FileLikeQueue()
+        # queue 32 bytes
+        q.write("*" * 7)
+        q.write("*" * 11)
+        q.write("*" * 5)
+        q.write("*" * 9)
+        q.close()
+        # read(-1) returns all, then ''
+        self.assertEqual(len(q.read(-1)), 32)
+        self.assertEqual(len(q.read(-1)), 0)
 
     def testStream(self):
         with WsgiDavTestServer(provider=self.provider):
             with Timing("testStream", self.SIZE):
                 with open(self.test_file, "rb") as f:
                     r = requests.put("http://127.0.0.1:8080/bar.txt", data=f)
-            self.assertEqual(r.status_code, 204)
-            self.assertEqual(os.path.getsize(self.target_path), self.SIZE)
+        self.assertEqual(r.status_code, 204)
+        self.assertEqual(os.path.getsize(self.target_path), self.SIZE)
 
     # def testStreamBlob(self):
     #     with WsgiDavTestServer(provider=self.provider):
